Added MainActivityViewModel as a shared ViewModel...
...This reduces redundant fields and ensures that badges are correctly displayed from all tabs
