+++ app/src/main/java/com/example/pomoplanner/ui/MainActivity.kt
@@ -9,15 +9,6 @@ import androidx.compose.foundation.interaction.Interaction
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.List
-import androidx.compose.material.icons.automirrored.outlined.List
-import androidx.compose.material.icons.filled.Notifications
-import androidx.compose.material.icons.filled.Person
-import androidx.compose.material.icons.filled.Settings
-import androidx.compose.material.icons.outlined.Notifications
-import androidx.compose.material.icons.outlined.Person
-import androidx.compose.material.icons.outlined.Settings
 import androidx.compose.material3.Badge
 import androidx.compose.material3.BadgedBox
 import androidx.compose.material3.Icon
@@ -30,59 +21,23 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.example.pomoplanner.model.DBHelper
 import com.example.pomoplanner.model.Profile
+import com.example.pomoplanner.ui.MainActivityViewModel.TabBarItem
 import com.example.pomoplanner.ui.theme.PomoPlannerTheme
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emptyFlow
-
-class TabBarItem(
-    val title: String,
-    val selectedIcon: ImageVector,
-    val unselectedIcon: ImageVector,
-    initialIsDisabled: Boolean = false,
-    initialBadgeAmount: Int? = null,
-) {
-    var isDisabled by mutableStateOf(initialIsDisabled)
-    var badgeAmount by mutableStateOf(initialBadgeAmount)
-}
-
-// setup of individual navigation tabs
-val profileTab = TabBarItem(
-    title = "Profile",
-    selectedIcon = Icons.Filled.Person,
-    unselectedIcon = Icons.Outlined.Person
-)
-val tasksTab = TabBarItem(
-    title = "Tasks",
-    selectedIcon = Icons.AutoMirrored.Filled.List,
-    unselectedIcon = Icons.AutoMirrored.Outlined.List,
-    true
-)
-val pomodoroTab = TabBarItem(
-    title = "Pomodoro",
-    selectedIcon = Icons.Filled.Notifications,
-    unselectedIcon = Icons.Outlined.Notifications,
-)
-val settingsTab = TabBarItem(
-    title = "Settings",
-    selectedIcon = Icons.Filled.Settings,
-    unselectedIcon = Icons.Outlined.Settings,
-)
-
-// creating list of navigation tabs
-val tabBarItems = listOf(profileTab, tasksTab, pomodoroTab, settingsTab)
 
 // application entry point
 class MainActivity : ComponentActivity() {
@@ -109,7 +64,11 @@ class MainActivity : ComponentActivity() {
 
 // app navigation controller
 @Composable
-private fun PomoPlannerApp() {
+private fun PomoPlannerApp(
+    mainActivityViewModel: MainActivityViewModel = viewModel(),
+) {
+    mainActivityViewModel.getCurrentTasks()
+
     // creating navController
     val navController = rememberNavController()
 
@@ -118,36 +77,36 @@ private fun PomoPlannerApp() {
         color = MaterialTheme.colorScheme.background
     ) {
         Scaffold(
-            bottomBar = { TabView(tabBarItems, navController) }
+            bottomBar = { TabView(mainActivityViewModel.tabBarItems, navController) }
         ) { innerPadding ->
             NavHost(
                 modifier = Modifier.padding(innerPadding),
                 navController = navController,
-                startDestination = profileTab.title
+                startDestination = mainActivityViewModel.profileTab.title
             ) {
-                composable(profileTab.title) {
+                composable(mainActivityViewModel.profileTab.title) {
                     BackHandler(true) {
                         // disable back navigating
                     }
-                    ProfileTab()
+                    ProfileTab(mainActivityViewModel)
                 }
-                composable(tasksTab.title) {
+                composable(mainActivityViewModel.tasksTab.title) {
                     BackHandler(true) {
                         // disable back navigating
                     }
-                    TasksTab()
+                    TasksTab(mainActivityViewModel)
                 }
-                composable(pomodoroTab.title) {
+                composable(mainActivityViewModel.pomodoroTab.title) {
                     BackHandler(true) {
                         // disable back navigating
                     }
-                    PomodoroTab()
+                    PomodoroTab(mainActivityViewModel)
                 }
-                composable(settingsTab.title) {
+                composable(mainActivityViewModel.settingsTab.title) {
                     BackHandler(true) {
                         // disable back navigating
                     }
-                    SettingsTab()
+                    SettingsTab(mainActivityViewModel)
                 }
             }
         }
