+++ app/src/main/java/com/example/pomoplanner/ui/PomodoroTab.kt
@@ -4,6 +4,8 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 
 @Composable
-fun PomodoroTab() {
+fun PomodoroTab(
+    mainActivityViewModel: MainActivityViewModel,
+) {
     Text("Pomodoro Tab")
 }
