+++ app/src/main/java/com/example/pomoplanner/ui/SettingsTab.kt
@@ -4,6 +4,8 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 
 @Composable
-fun SettingsTab() {
+fun SettingsTab(
+    mainActivityViewModel: MainActivityViewModel,
+) {
     Text("Settings Tab")
 }
